fix a bug where config fiels would not be overriden

+++ link_dotfiles.sh
@@ -4,15 +4,30 @@ echo 'Setting up Symlinks.'
 ln -sf /usr/local/bin/python3 /usr/local/bin/python
 ln -sf /usr/local/bin/pip3 /usr/local/bin/pip
 ln -sf "$(pwd)/git/gitignore_global" ~/.gitignore_global
-ln -sf "$(pwd)/git/gitconfig" ~/.gitconfig
 ln -sf "$(pwd)/tmux/tmux.conf" ~/.tmux.conf
 ln -sf "$(pwd)/zsh/zshrc" ~/.zshrc
 
+echo "Linking and setting up nvim"
 mkdir -p ~/.config/nvim/
 ln -sf "$(pwd)/nvim/init.vim" ~/.config/nvim/init.vim
-
-echo "Installing Neovim Plugin Manager"
 nvim --headless +PlugInstall +qa
+
+if [ -f "$(pwd)/ssh/config" ]; then
+    echo "Linking ssh config"
+    mkdir -p "$HOME/.ssh/"
+    ln -sf "$(pwd)/ssh/config" ~/.ssh/config
+else
+    echo "No ssh config file to link found at $(pwd)/ssh/config"
+fi
+
+
+
+if [ -f "$(pwd)/git/gitconfig" ]; then
+    echo "Linking git config"
+    ln -sf "$(pwd)/git/gitconfig" ~/.gitconfig
+else
+    echo "No ssh config file to link found at $(pwd)/git/gitconfig"
+fi
 
 
 
@@ -27,6 +42,7 @@ ln -sf "$(pwd)/vscode/snippets/python.json" "$CODE_SNIPPETS_DIR/python.json"
 
 
 if [ -d "/Applications/Karabiner-Elements.app" ]; then
+    echo "Karabiner installation found."
     if  xattr "/Applications/Karabiner-Elements.app" | grep quarantine ; then
         echo "adding gatekeeper exception for karabiner"
         sudo xattr -r -d com.apple.quarantine "/Applications/Karabiner-Elements.app"
